refactor: 변경된 경로에 맞게 import문 수정 #595
