+++ frontend/src/common/apis/getMineMentoring.ts
@@ -2,7 +2,7 @@ import { API_ENDPOINTS } from '../constants/apiEndpoints';
 
 import { apiClient } from './apiClient';
 
-import type { MentoringResponse } from '../../pages/detail/types/MentoringResponse';
+import type { MentoringResponse } from '../types/MentoringResponse';
 
 export const getMineMentoring = () => {
   return apiClient.get<MentoringResponse>({
